Moved `TileInfo` from `Features` to `Utils` (khanshoaib3/stardew-access#112).
