+++ stardew-access/Patches/MiscPatches/Game1Patch.cs
@@ -1,5 +1,4 @@
 using Microsoft.Xna.Framework;
-using stardew_access.Features;
 using stardew_access.Utils;
 using StardewModdingAPI;
 using StardewValley;
